// types/database.types.ts
export type Database = {
  public: {
    Tables: {
      ingredients: {
        Row: {
          id: number
          name: string
          recipe_id: number | null
        }
        Insert: {
          id?: never
          name: string
          recipe_id?: number | null
        }
        Update: {
          id?: never
          name?: string
          recipe_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "ingredients_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
        ]
      }
      instructions: {
        Row: {
          description: string
          id: number
          recipe_id: number | null
          step_number: number
        }
        Insert: {
          description: string
          id?: never
          recipe_id?: number | null
          step_number: number
        }
        Update: {
          description?: string
          id?: never
          recipe_id?: number | null
          step_number?: number
        }
        Relationships: [
          {
            foreignKeyName: "instructions_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
        ]
      }
      recipe_upvotes: {
        Row: {
          created_at: string | null
          id: number
          recipe_id: number | null
          user_id: number | null
        }
        Insert: {
          created_at?: string | null
          id?: never
          recipe_id?: number | null
          user_id?: number | null
        }
        Update: {
          created_at?: string | null
          id?: never
          recipe_id?: number | null
          user_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "recipe_upvotes_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "recipe_upvotes_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      recipes: {
        Row: {
          calories: number | null
          chef: string | null
          cook_time: string | null
          created_at: string | null
          description: string | null
          difficulty: string | null
          id: number
          image_url: string | null
          is_public: boolean | null
          name: string
          servings: number | null
          user_id: number | null
        }
        Insert: {
          calories?: number | null
          chef?: string | null
          cook_time?: string | null
          created_at?: string | null
          description?: string | null
          difficulty?: string | null
          id?: never
          image_url?: string | null
          is_public?: boolean | null
          name: string
          servings?: number | null
          user_id?: number | null
        }
        Update: {
          calories?: number | null
          chef?: string | null
          cook_time?: string | null
          created_at?: string | null
          description?: string | null
          difficulty?: string | null
          id?: never
          image_url?: string | null
          is_public?: boolean | null
          name?: string
          servings?: number | null
          user_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "recipes_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      users: {
        Row: {
          created_at: string | null
          email: string
          full_name: string
          id: number
          user_id: string
        }
        Insert: {
          created_at?: string | null
          email: string
          full_name: string
          id?: never
          user_id: string
        }
        Update: {
          created_at?: string | null
          email?: string
          full_name?: string
          id?: never
          user_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never